Move options struct file to config package
