+++ cmd/manager/main.go
@@ -12,7 +12,7 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 	"k8s.io/client-go/rest"
 
-	dv_config "github.com/app-sre/deployment-validation-operator/config"
+	dvConfig "github.com/app-sre/deployment-validation-operator/config"
 	"github.com/app-sre/deployment-validation-operator/pkg/apis"
 	"github.com/app-sre/deployment-validation-operator/pkg/controller"
 	dvo_prom "github.com/app-sre/deployment-validation-operator/pkg/prometheus"
@@ -39,9 +39,9 @@ const operatorNameEnvVar = "OPERATOR_NAME"
 
 func main() {
 	// Make sure the operator name is what we want
-	os.Setenv(operatorNameEnvVar, dv_config.OperatorName)
+	os.Setenv(operatorNameEnvVar, dvConfig.OperatorName)
 
-	opts := options{
+	opts := dvConfig.Options{
 		MetricsPort: 8383,
 		MetricsPath: "metrics",
 		ProbeAddr:   ":8081",
@@ -76,7 +76,7 @@ func main() {
 	}
 }
 
-func setupManager(log logr.Logger, opts options) (manager.Manager, error) {
+func setupManager(log logr.Logger, opts dvConfig.Options) (manager.Manager, error) {
 	logVersion(log)
 
 	log.Info("Load KubeConfig")
@@ -185,7 +185,7 @@ func initializeScheme() (*k8sruntime.Scheme, error) {
 
 var errWatchNamespaceNotSet = errors.New("'WatchNamespace' not set")
 
-func getManagerOptions(scheme *k8sruntime.Scheme, opts options) (manager.Options, error) {
+func getManagerOptions(scheme *k8sruntime.Scheme, opts dvConfig.Options) (manager.Options, error) {
 	ns, ok := opts.GetWatchNamespace()
 	if !ok {
 		return manager.Options{}, errWatchNamespaceNotSet
